services/api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "benefits-opportunity-api"
version = "0.1.0"
description = "FastAPI backend for the Benefits Opportunity Engine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.1",
  "beautifulsoup4>=4.13.3",
  "dramatiq>=1.17.1",
  "email-validator>=2.2.0",
  "fastapi>=0.115.11",
  "itsdangerous>=2.2.0",
  "httpx>=0.28.1",
  "openai>=1.66.5",
  "psycopg[binary]>=3.2.6",
  "pydantic-settings>=2.8.1",
  "python-multipart>=0.0.20",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.39",
  "sentry-sdk[fastapi]>=2.23.0",
  "boto3>=1.37.11",
  "pypdf>=5.4.0",
  "pytest>=8.3.5",
  "uvicorn[standard]>=0.34.0"
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
